Clean up ANSI warnings
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1407 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/nethack.d/pri.c
@@ -427,7 +427,7 @@ extern char *eos();
 	(void) strcat(newbot, hu_stat[u.uhs]);
 	if(flags.time)
 	    (void) sprintf(eos(newbot), "  %ld", moves);
-	if(strlen(newbot) >= COLNO) {
+	if(strlen(newbot) >= (unsigned)COLNO) {
 		register char *bp0, *bp1;
 		bp0 = bp1 = newbot;
 		do {
